requirements/base.txt
asn1crypto==0.24.0

idna==2.7
pbr==4.2.0
pipsi==0.9
ply==3.4
stevedore==1.29.0
Pillow==4.2.1
whitenoise==3.3.1
pyOpenSSL==18.0.0
amqp==2.3.2
appdirs==1.4.3
billiard==3.5.0.4
cairocffi==0.8.1

configparser==3.5.0
cssselect2==0.2.1
dj-database-url==0.5.0
django-appconf==1.0.2
django-extensions==1.7.5
django-redis-cache==1.7.1
django-redis==4.8.0
django-render-block==0.5
django-summernote==0.8.11.6
django-templated-email==2.2.0
django-user-sessions==1.7.1
enum34==1.1.6
geoip2==2.4.2
html5lib==1.0.1
ipaddress==1.0.18
mysqlclient==2.0.3

pycodestyle==2.3.1
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.6.1
python-decouple==3.0
python-magic==0.4.15
pytz==2017.2
pytzdata==2017.2
rcssmin==1.0.6
redis==2.10.6
requests==2.20
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.3
tinycss2==0.6.1
tzlocal==1.4
webencodings==0.5.1
xlwt==1.3.0
CairoSVG==1.0.22
django_compressor==2.2
Django==2.2.4
Pyphen==0.9.4
WeasyPrint==0.42
django-file-resubmit==0.5.2
maxminddb==2.0.3
boto3==1.16.59
django-storages==1.11.1
stripe==2.59.0